Show what each endpoint really answers with
Browsing shows the packages on sale, not just listings. Only a listing
sold by the day offers dates to choose. Deactivating answers with a
deactivated listing, and reactivating with a live one. The undated
package example no longer sends a date the endpoint would ignore.

The examples file is split by audience — the public endpoints, the
shared endpoint shape, and the admin ones — and a stray comment is back
on the declaration it describes.

Raised by Codex on #1975.
