+++ src/shared/admin-api-example.ts
@@ -6,7 +6,6 @@
  * output, so a shape change will break the test and force an update.
  */
 
-import * as v from "valibot";
 import {
   type CreateListingBody,
   type DeleteListingBody,
@@ -23,18 +22,14 @@ import type {
   DeleteHolidayBody,
   UpdateHolidayBody,
 } from "#routes/admin/api-holidays.ts";
-import type { PublicListing } from "#routes/api/public-listing.ts";
-import { PackageChildrenSchema } from "#routes/api/request-schemas.ts";
-import {
-  API_AVAILABILITY_EXAMPLE_JSON,
-  API_BOOK_PAID_EXAMPLE_JSON,
-  API_EXAMPLE_LISTING,
-  API_LIST_EXAMPLE_JSON,
-  API_SINGLE_EXAMPLE_JSON,
-} from "#shared/api-example.ts";
+import { API_EXAMPLE_LISTING } from "#shared/api-example.ts";
 import { listingCatalogFields } from "#shared/catalog-fields/fields.ts";
 import { VALID_DAY_NAMES } from "#shared/day-names.ts";
 import type { AdminListing } from "#shared/types.ts";
+import { type EndpointDoc, json } from "./admin-api-example/endpoint-doc.ts";
+
+export type { EndpointDoc } from "./admin-api-example/endpoint-doc.ts";
+export { PUBLIC_API_ENDPOINTS } from "./admin-api-example/public.ts";
 
 /** The example listing exactly as the admin endpoints answer with it: the
  * stored fields, plus the ids of the groups it is in. The example is in none. */
@@ -132,131 +127,8 @@ const ADMIN_API_HOLIDAY_DELETE_BODY = {
   confirm_identifier: "Christmas",
 } satisfies DeleteHolidayBody;
 
-// =============================================================================
-// Endpoint documentation entries
-// =============================================================================
-
-/** A documented API endpoint with example request and response */
-export type EndpointDoc = {
-  method: string;
-  path: string;
-  description: string;
-  request?: string;
-  response: string;
-};
-
 /** The booking window a listing gets when its create body says nothing. */
 const LISTING_DEFAULT_DAYS_AFTER = listingCatalogFields.maximumDaysAfter[4];
-
-const json = (data: unknown): string => JSON.stringify(data, null, 2);
-
-/** The add-on offered under the example package's "Tent Pitch" member. A member
- * that offers a child is published with it, so the booking example can choose
- * it by slug. */
-const PACKAGE_EXAMPLE_CHILD = {
-  canPayMore: false,
-  customisableDays: false,
-  date: null,
-  description: "A duvet and pillows for the tent.",
-  fields: "email",
-  imageAltText: null,
-  imageUrl: null,
-  isClosed: false,
-  isSoldOut: false,
-  listingType: "standard",
-  location: null,
-  maxPrice: 1200,
-  maxPurchasable: 5,
-  name: "Extra Bedding",
-  nonTransferable: false,
-  purchaseOnly: false,
-  slug: "extra-bedding",
-  unitPrice: 1200,
-} satisfies PublicListing;
-
-/** The package-book example's `children`, parsed through the LIVE request
- * schema ({@link PackageChildrenSchema}) — a drifted example is a build-time
- * parse error, so the docs can never show a body the endpoint rejects. */
-const PACKAGE_BOOK_CHILDREN_EXAMPLE = v.parse(PackageChildrenSchema, [
-  { parent: "tent-pitch", quantity: 1, slug: "extra-bedding" },
-]);
-
-export const PUBLIC_API_ENDPOINTS: EndpointDoc[] = [
-  {
-    description: "List all active, non-hidden listings",
-    method: "GET",
-    path: "/api/listings",
-    response: API_LIST_EXAMPLE_JSON,
-  },
-  {
-    description: "Get a single listing by slug",
-    method: "GET",
-    path: "/api/listings/:slug",
-    response: API_SINGLE_EXAMPLE_JSON,
-  },
-  {
-    description:
-      "Check if spots are available (optional query: quantity, date)",
-    method: "GET",
-    path: "/api/listings/:slug/availability",
-    response: API_AVAILABILITY_EXAMPLE_JSON,
-  },
-  {
-    description: "Create a booking",
-    method: "POST",
-    path: "/api/listings/:slug/book",
-    request: json({
-      email: "alice@example.com",
-      name: "Alice Smith",
-      quantity: 2,
-    }),
-    // The listing has a price, so booking it answers with somewhere to pay.
-    response: API_BOOK_PAID_EXAMPLE_JSON,
-  },
-  {
-    description:
-      "Get a package bundle by slug: its whole-bundle price (per day count for customisable-days bundles), capacity, dates, and members with their required children",
-    method: "GET",
-    path: "/api/packages/:slug",
-    response: json({
-      package: {
-        description: "Two nights' camping with firepit hire",
-        fields: "email,phone",
-        maxPurchasable: 5,
-        members: [
-          {
-            // The add-on the booking example below chooses for this member.
-            children: [PACKAGE_EXAMPLE_CHILD],
-            name: "Tent Pitch",
-            quantity: 1,
-            slug: "tent-pitch",
-          },
-          { name: "Firepit", quantity: 1, slug: "firepit" },
-        ],
-        name: "Camping Weekend",
-        priceMinor: 5500,
-        slug: "camping-weekend",
-      },
-    }),
-  },
-  {
-    description:
-      "Book whole package bundles (optional: date for dated bundles, dayCount for customisable ones, children choosing each parent member's add-ons)",
-    method: "POST",
-    path: "/api/packages/:slug/book",
-    request: json({
-      children: PACKAGE_BOOK_CHILDREN_EXAMPLE,
-      date: "2025-08-20",
-      email: "alice@example.com",
-      name: "Alice Smith",
-      // The package asks for email and phone, so a booking must give both.
-      phone: "+447700900123",
-      quantity: 1,
-    }),
-    // The bundle costs money, so booking it answers with somewhere to pay.
-    response: API_BOOK_PAID_EXAMPLE_JSON,
-  },
-];
 
 /** The five standard admin-CRUD doc entries for a resource. Descriptions are
  *  passed in (they carry per-resource wording — "an listing", the holiday
@@ -366,13 +238,17 @@ export const ADMIN_API_ENDPOINTS: EndpointDoc[] = [
     description: "Deactivate an listing",
     method: "POST",
     path: "/api/admin/listings/:listingId/deactivate",
-    response: json({ listing: ADMIN_API_EXAMPLE_ADMIN_LISTING }),
+    response: json({
+      listing: { ...ADMIN_API_EXAMPLE_ADMIN_LISTING, active: false },
+    }),
   },
   {
     description: "Reactivate a deactivated listing",
     method: "POST",
     path: "/api/admin/listings/:listingId/reactivate",
-    response: json({ listing: ADMIN_API_EXAMPLE_ADMIN_LISTING }),
+    response: json({
+      listing: { ...ADMIN_API_EXAMPLE_ADMIN_LISTING, active: true },
+    }),
   },
   ...crudDocs({
     createBody: ADMIN_API_GROUP_CREATE_BODY,
